requirements.txt
streamlit
pandas
openai
numpy
python-dotenv #to fetch key from .env
Pillow #pixel data to vector data
